requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# LLM
openai==1.28.0

# Embeddings
sentence-transformers==3.0.0
torch==2.3.0
numpy==1.26.4

# Vector DB
qdrant-client==1.9.1

# PDF parsing
PyMuPDF==1.24.5
pdfplumber==0.11.1

# Database
sqlalchemy==2.0.30
aiosqlite==0.20.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
httpx==0.27.0
aiofiles==23.2.1
tiktoken==0.7.0

# Export
python-docx==1.1.2
reportlab==4.2.0

# Background tasks
anyio==4.4.0
